requirements.txt
Brlapi==0.6.1
<<<<<<< HEAD
Django==2.1.4
=======
Django==2.0
>>>>>>> dennis_dev
Mako==0.9.1
MarkupSafe==0.18
apturl==0.5.2ubuntu4.1
chardet==2.2.1
checkbox-ng==0.3
checkbox-support==0.2
colorama==0.2.5
command-not-found==0.3
defer==1.0.6
feedparser==5.1.3
friends==0.1
html5lib==0.999
httplib2==0.8
language-selector==0.1
louis==2.5.3
lxml==3.3.3
oauthlib==0.6.1
onboard==1.0.1
oneconf==0.3.7.14.04.1
piston-mini-client==0.7.5
plainbox==0.5.3
pycrypto==2.6.1
pycurl==7.19.3
pygobject==3.12.0
pyparsing==2.0.1
python-apt==0.9.3.5ubuntu3
python-debian==0.1.21-nmu2ubuntu2
<<<<<<< HEAD
pytz==2018.7
pyxdg==0.25
requests==2.2.1
six==1.5.2
software-center-aptd-plugins==0.0.0
ubuntu-drivers-common==0.0.0
ufw==0.34-rc-0ubuntu2
unattended-upgrades==0.1
unity-scope-audacious==0.1
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-clementine==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-gdrive==0.7
unity-scope-gmusicbrowser==0.1
unity-scope-gourmet==0.1
unity-scope-guayadeque==0.1
unity-scope-manpages==0.1
unity-scope-musique==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
urllib3==1.7.1
usb-creator==0.2.23
virtualenv==16.1.0
wheel==0.24.0
xdiagnose==3.6.3build2
xkit==0.0.0
=======
amqp==2.3.2
asn1crypto==0.24.0
attrs==18.2.0
autobahn==18.11.2
Automat==0.7.0
bcrypt==3.1.4
billiard==3.5.0.5
celery==4.2.1
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
constantly==15.1.0
cryptography==2.4.2
dateparser==0.7.0
Django==2.0.9
django-js-asset==1.1.0
django-mptt==0.9.1
django-redis==4.10.0
djangorestframework==3.9.0
fabric==2.4.0
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
invoke==1.2.0
jsonfield==2.0.2
kombu==4.2.1
mysqlclient==1.3.14
paramiko==2.4.2
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.19
PyHamcrest==1.9.0
PyNaCl==1.3.0
pyOpenSSL==18.0.0
python-binance==0.7.0
python-dateutil==2.7.5
pytz==2018.7
redis==3.0.1
regex==2018.11.22
requests==2.20.1
service-identity==17.0.0
six==1.11.0
Twisted==18.9.0
txaio==18.8.1
tzlocal==1.5.1
urllib3==1.24.1
vine==1.1.4
zope.interface==4.6.0
>>>>>>> dennis_dev
=======
pytz==2018.7
pyxdg==0.25
requests==2.2.1
six==1.5.2
software-center-aptd-plugins==0.0.0
ubuntu-drivers-common==0.0.0
ufw==0.34-rc-0ubuntu2
unattended-upgrades==0.1
unity-scope-audacious==0.1
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-clementine==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-gdrive==0.7
unity-scope-gmusicbrowser==0.1
unity-scope-gourmet==0.1
unity-scope-guayadeque==0.1
unity-scope-manpages==0.1
unity-scope-musique==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
urllib3==1.7.1
usb-creator==0.2.23
virtualenv==16.1.0
wheel==0.24.0
xdiagnose==3.6.3build2
xkit==0.0.0
>>>>>>> dennis_dev
